:recycle: change it to the uin128 in function

+++ src/PricingModules/AbstractPricingModule.sol
@@ -223,7 +223,7 @@ abstract contract PricingModule is Ownable {
      * @param amount The amount of tokens
      * @dev Unsafe cast to uint128, meaning it is assumed no more than 10**(20+decimals) tokens will ever be deposited
      */
-    function processWithdrawal(address asset, uint256 amount) external virtual onlyMainReg {
-        exposure[asset].exposure -= uint128(amount);
+    function processWithdrawal(address asset, uint128 amount) external virtual onlyMainReg {
+        exposure[asset].exposure -= amount;
     }
 }
